Fixed threading/synchronization stuff for Linux

+++ src/shadow/natives/Pointer.c
@@ -13,5 +13,5 @@ void __shadow_natives__Pointer_free(shadow_Pointer_t* _this, void* ptr)
 
 void __printRef(shadow_Object_t* object)
 {
-    printf("Reference count: %llu\n", object->ref_count);
+    printf("Reference count: %llu\n", (unsigned long long)object->ref_count);
 }
